requirements.txt
asgiref==3.3.1
Django==3.1.5
django-summernote==0.8.11.6
djangorestframework==3.12.2
Pillow==7.0.0
pytz==2020.5
sqlparse==0.4.1
